+++ mcu/attiny-hal/src/lib.rs
@@ -10,12 +10,17 @@ compile_error!(
     Please select one of the following
 
     * attiny85
+    * attiny88
     "
 );
 
 /// Reexport of `attiny85` from `avr-device`
 #[cfg(feature = "attiny85")]
 pub use avr_device::attiny85 as pac;
+
+/// Reexport of `attiny88` from `avr-device`
+#[cfg(feature = "attiny88")]
+pub use avr_device::attiny88 as pac;
 
 /// See [`avr_device::entry`](https://docs.rs/avr-device/latest/avr_device/attr.entry.html).
 #[cfg(feature = "rt")]
@@ -39,5 +44,12 @@ pub struct Attiny;
 macro_rules! pins {
     ($p:expr) => {
         $crate::Pins::new($p.PORTB)
+    };
+}
+#[cfg(feature = "attiny88")]
+#[macro_export]
+macro_rules! pins {
+    ($p:expr) => {
+        $crate::Pins::new($p.PORTA, $p.PORTB, $p.PORTC, $p.PORTD)
     };
 }
